+++ source/eogmaneo/ComputeSystem.h
@@ -36,12 +36,10 @@ namespace eogmaneo {
 		
 		friend class Layer;
 		friend class Hierarchy;
+		friend class Agent;
+		friend class KMeansEncoder;
         friend class ImageEncoder;
-		friend class RandomEncoder;
-        friend class GaborEncoder;
-		friend class CornerEncoder;
-		friend class LocalRegressor;
-		
+
 		friend std::vector<float> whiten(const std::vector<float> &src, int width, int radius, float strength, ComputeSystem &cs, int chunkSize);
         friend std::vector<float> sobel(const std::vector<float> &src, int width, float clip, ComputeSystem &cs, int chunkSize);
     };